Refactor yaml key sorting utility to a separate utils package
